==> notebooks/1.extraccion.ipynb <==
# 1. Extracción a través de Apify

### El proceso es muy sencillo. Se llama a la API con un agente preentrenado para el scrapeo de Metacritic y después se llama a la base datos scrapeada y se importa como CSV

#%pip install apify-client

from apify_client import ApifyClient     # importamos cliente de API externa
import requests as req
import pandas as pd

with open('apify.txt', 'r') as archivo:
    token = archivo.read()

apify_client = ApifyClient(token)   # inicializamos cliente de API externa

actor_call = apify_client.actor('epctex/metacritic-scraper').call()

apify_dataset_url = 'https://api.apify.com/v2/datasets/WRSzhAp6IzSGw9xh6/items?format=json'  # URL del dataset final desde API externa

response = req.get(apify_dataset_url)

if response.status_code == 200:
    dataset_data = response.json()         # obtenemos datos del dataset

df = pd.DataFrame(dataset_data)        # creamos dataframe con los datos del dataset

df.info()                            # mostramos información del dataframe

df.to_csv('metacritic.csv', index=False)  # guardamos dataframe en archivo csv
