==> notebooks/2.transformacion.ipynb <==
# 2. Transformación de los datos con Pandas

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import ast
from src.funciones import *

df = pd.read_csv('..\\Proyecto_Final\\data\\metacritic.csv')  # importamos el dataset

### Realizamos una exploración de los datos para después proceder a su transformación y limpieza. En primer lugar, manejaremos nulos y duplicados, para luego pasar a las columnas que queremos transformar.

df.info()  # exploramos el dataset, obervamos que hay algunos nulos

df.shape  # observamos que hay 12703 registros y 14 columnas

df.head()  # observamos las primeras 5 filas del dataset

df['type'].unique()  # solo hay un un valor en la columna type, por lo que no aporta información. La eliminamos.

df = df.drop(['type'], axis=1)  # eliminamos la columna type

df['awards'].unique()   # la columna awards tampoco nos aporta valor, por lo que también la eliminamos

df = df.drop(['awards'], axis=1)  # eliminamos la columna awards

df.isna().sum()  # al observar los nulos, vemos que se concentran su mayoría en las columnas 'release date' y 'summary'

df = df.dropna() #prescindimos de los nulos ya que son muy pocos registros

df.shape  # se han elimnado 99 registros

df = df.reindex(columns= ['metascore','userscore', 'title', 'genres', 'section', 'publisherName', 'platforms', 'releaseDate', 'summary', 'userReviewsSummary','url','publisherUrl'])
df.head()    # hacemos un reindex de las columnas para que nos sea más facil visualizar los datos que nos interesan

### COLUMNA METASCORE

df['metascore'] = df['metascore'].astype(int) # convertimos a enteros los valores de metascore, ya que no tienen decimales

df = df.sort_values(by='metascore', ascending=False) # ordenamos el dataframe por los valores de la columna metascore

df.head()

### COLUMNA GENRES

df.genres.unique() # hay muchos valores para la columna Genres, por lo que intentaremos unificarlos para facilitar los filtros

df['genres'] = df['genres'].apply(lambda x: ', '.join(ast.literal_eval(x))) # usamos Abstract Syntax Trees para sacar los valores de la lista

# Aplicamos un mapeo para unificar los valores de la columna genres

genre_mapping = {
    "2D Beat-'Em-Up": 'Fighting',
    '2D Fighting': 'Fighting',
    '2D Platformer': 'Platformer',
    "3D Beat-'Em-Up": 'Fighting',
    '3D Fighting': 'Fighting',
    '3D Platformer': 'Platformer',
    '4X Strategy': 'Strategy',
    'Action': 'Action',
    'Action Adventure': 'Adventure',
    'Action Puzzle': 'Puzzle',
    'Action RPG': 'RPG',
    'Adventure': 'Adventure',
    'Aircraft Combat Sim': 'Simulation',
    'Aircraft Sim': 'Simulation',
    'Application': 'Others',
    'Arcade': 'Arcade',
    'Arcade Racing': 'Racing',
    'Artillery': 'Strategy',
    'Athletics': 'Sports',
    'Auto Racing': 'Racing',
    'Auto Racing Sim': 'Racing',
    'Baseball': 'Sports',
    'Baseball Sim': 'Sports',
    'Basketball': 'Sports',
    'Basketball Sim': 'Sports',
    'Biking': 'Sports',
    'Billiards': 'Sports',
    'Board': 'Puzzle',
    'Bowling': 'Sports',
    'Card Battle': 'Strategy',
    'Combat Sport': 'Fighting',
    'Command RTS': 'Strategy',
    'Compilation': 'Others',
    'Cricket': 'Sports',
    'Dancing': 'Music',
    'Defense': 'Strategy',
    'Edutainment': 'Education',
    'Exercise': 'Sports',
    'FPS': 'Shooter',
    'First-Person Adventure': 'Adventure',
    'Fishing': 'Simulation',
    'Football': 'Sports',
    'Football Sim': 'Sports',
    'Future Racing': 'Racing',
    'Future Sport': 'Sports',
    'Gambling': 'Strategy',
    'Golf': 'Sports',
    'Golf Sim': 'Sports',
    'Hidden Object': 'Puzzle',
    'Hockey': 'Sports',
    'Hockey Sim': 'Sports',
    "Horizontal Shoot-'Em-Up": 'Shooter',
    'Horse Racing': 'Sports',
    'Hunting': 'Simulation',
    'Individual Sports': 'Sports',
    'JRPG': 'RPG',
    'Light Gun': 'Shooter',
    'Linear Action Adventure': 'Adventure',
    'Logic Puzzle': 'Puzzle',
    'MMORPG': 'RPG',
    'MOBA': 'Strategy',
    'Management': 'Strategy',
    'Marine Combat Sim': 'Simulation',
    'Marine Sim': 'Simulation',
    'Matching Puzzle': 'Puzzle',
    'Metroidvania': 'Platformer',
    'Miscellaneous': 'Others',
    'Open-World Action': 'Action',
    'Party': 'Casual',
    'Pinball': 'Arcade',
    'Point-and-Click': 'Adventure',
    'Puzzle': 'Puzzle',
    'RPG': 'RPG',
    'Racing': 'Racing',
    'Racing Sim': 'Racing',
    'Rail Shooter': 'Shooter',
    'Real-Time Strategy': 'Strategy',
    'Real-Time Tactics': 'Strategy',
    'Rhythm': 'Music',
    'Roguelike': 'RPG',
    'Rugby': 'Sports',
    'Sandbox': 'Simulation',
    'Simulation': 'Simulation',
    'Skating': 'Sports',
    'Skiing': 'Sports',
    'Soccer': 'Sports',
    'Soccer Management': 'Strategy',
    'Soccer Sim': 'Sports',
    'Space Combat Sim': 'Simulation',
    'Space Sim': 'Simulation',
    'Sports': 'Sports',
    'Stacking Puzzle': 'Puzzle',
    'Strategy': 'Strategy',
    'Surfing': 'Sports',
    'Survival': 'Adventure',
    'Tactical FPS': 'Shooter',
    'Tactical Third Person Shooter': 'Shooter',
    'Team Sports': 'Sports',
    'Tennis': 'Sports',
    'Text Adventure': 'Adventure',
    'Third Person Shooter': 'Shooter',
    'Third-Person Adventure': 'Adventure',
    "Top-Down Shoot-'Em-Up": 'Shooter',
    'Train Sim': 'Simulation',
    'Trainer RPG': 'RPG',
    'Trivia': 'Puzzle',
    'Turn-Based Strategy': 'Strategy',
    'Turn-Based Tactics': 'Strategy',
    'Tycoon': 'Strategy',
    'Vehicle Combat Sim': 'Simulation',
    'Vehicle Sim': 'Simulation',
    "Vertical Shoot-'Em-Up": 'Shooter',
    'Virtual Career': 'Simulation',
    'Virtual Life': 'Simulation',
    'Virtual Pet': 'Casual',
    'Visual Novel': 'Adventure',
    'Volleyball': 'Sports',
    'Western RPG': 'RPG',
    'Wrestling': 'Sports'
}



# Aplicación de la función a toda la columna de géneros en tu DataFrame
df['genres'] = df['genres'].apply(apply_genre_mapping)


df.head()

### COLUMNA REALEASE DATE

df['releaseDate'] = pd.to_datetime(df['releaseDate']).dt.strftime('%d-%m-%Y')

 # convertimos la columna a datetime y cambiamos el formato

### COLUMNA PLATFORMS

df['platforms'].unique() #observamos que los valores de la columna platforms están en listas

df['platforms']  = df['platforms'].apply(lambda x: ', '.join(ast.literal_eval(x)))  
# usamos el módulo ast para evaluar las listas en las cadenas y sacar los elementos

### COLUMNA USER REVIEW SUMMARY

df['userReviewsSummary'].unique()

df['userReviewsSummary'] = df['userReviewsSummary'].apply(ast.literal_eval)  # obtenemos los valores de la lista

df['reviewCount'] = df['userReviewsSummary'].apply(lambda x: x.get('reviewCount'))  # para cada valor, creamos una columna nueva
df['positive'] = df['userReviewsSummary'].apply(lambda x: x.get('positive'))
df['neutral'] = df['userReviewsSummary'].apply(lambda x: x.get('neutral'))
df['negative'] = df['userReviewsSummary'].apply(lambda x: x.get('negative'))

df.drop(['userReviewsSummary'], axis=1, inplace=True)  # eliminamos la columna userReviewsSummary, ya que tenemos la información en las nuevas columnas

### ÚLTIMOS PASOS: REINDEX, REDUCCIÓN Y CATEGORIZACIÓN DE COLUMNAS NUMÉRICAS

# vamos a renombrar columnas para que sean más claras

columnas = {'publisherName': 'publisher', 'releaseDate': 'release','publisherUrl': 'publisher_url','reviewCount': 'reviews', 'url': 'game_url'}
df.rename(columns=columnas, inplace=True)

df.columns = df.columns.str.capitalize()  # capitalizamos las columnas

df = df.reset_index(drop=True)  # reseteamos el índice

df.info() #vamos a proceder a reducir el tamaño de los datos para que sea más fácil trabajar con ellos

int_columns = df.select_dtypes(include=['int64']).columns  # primero reducimos el tamaño de las columnas enteras
df[int_columns] = df[int_columns].astype('int32')

df['Userscore'] = df['Userscore'].astype('float32')  # reducimos el tamaño de la columna Userscore

df = df.apply(categorize)  # utilizamos unas función para categorizar

df.info(memory_usage= 'deep')  # observamos que el tamaño del dataset se ha reducido considerablemente

### EXPORTAMOS EL CSV TRANSFORMADO

df.to_csv('..\\Proyecto_Final\\data\\metacritic_transform.csv', index=False)  # exportamos el dataset transformado
