==> notebooks/3.carga.ipynb <==
# 3. Carga a DB de SQL

import mysql.connector as conn
import pandas as pd

### En primer lugar, se crea la base y las tablas a través de MySQL Connector

with open('password.txt', 'r') as archivo:   # Creación de la conexión a la base de datos
    contraseña = archivo.read()

mydb = conn.connect(
  host="localhost",
  user="root",
  password= contraseña
)

mycursor = mydb.cursor()

mycursor.execute("CREATE DATABASE metacritic")

mydb = conn.connect(   # Creación de las tablas de la base de datos
  host="localhost",
  user="root",
  password= contraseña,
  database="metacritic"
)

mycursor = mydb.cursor()

mycursor.execute('drop table if exists games;')



query = '''


create table games(

            
        id int primary key not null,
        Metascore int,
        Userscore float,
        Title varchar(100),
        Genres varchar(100),
        Section varchar(50),
        Publisher varchar(50),
        Platforms varchar(500),
        Release datetime,
        Summary varchar(10000),
        Game_url varchar(10000),
        Publisher_url varchar(10000),
        Reviews int,
        Positive int,
        Neutral int,
        Negative int
);

'''

### En segundo lugar, se utiliza SQL Alchemy para rellenar la tabla con nuestros datos directamente a la basa de datos

from sqlalchemy import create_engine           # Creación del cursor para la conexión a la base de datos

str_conn = f'mysql+pymysql://root:{contraseña}@localhost:3306/metacritic'
cursor = create_engine(str_conn)

pd.read_csv('..\\Proyecto_Final\\data\\metacritic_transform.csv').to_sql('games', con = cursor, if_exists='append', index=False)

# Se rellenan y se exportan las tablas de la base de datos
