==> notebooks/4.traduccion.ipynb <==
# 4. Traducción de los datos

### Los datos recibidos están en inglés. Es interesante tener la base de datos tanto en inglés como en español para posibles usos futuros, por lo que haremos una pequeña transformación de los datos


#%pip install googletrans==4.0.0-rc1

import pandas as pd                   
from googletrans import Translator  # utilizaremos la librería googletrans para traducir los textos al español
from src.funciones import * # importamos las funciones que hemos creado en el archivo funciones.py

df = pd.read_csv('..\\Proyecto_Final\\data\\metacritic_transform.csv')

df.columns

df['Summary'] = df['Summary'].apply(traducir_texto, destino='es') 
  
 # traducimos los textos de la columna Summary utilizando una función para traducir texto

df.rename(columns={'Summary':'Resumen', 'Title':'Título', 'Genres':'Género','Section':'Sección', 'Publisher':'Distribuidor','Platforms':'Plataformas','Release':'Lanzamiento','Game_url':'Juego_url','Publisher_url':'Distribuidor_url','Reviews':'Reseñas','Positive':'Positiva','Negative':'Negativa'}, inplace=True)

# renombramos las columnas para que tengan nombres en español

map = {
'Action': 'Acción',
'Fighting':'Lucha',
'Platformer':'Plataformas',
'Sports':'Deportes',
'Shooter':'Shooter',
'RPG':'RPG',
'Others':'Otros',
'Racing':'Carreras',
'Adventure':'Aventura',
'Strategy':'Estrategia',
'Music':'Música',
'Arcade':'Arcade',
'Simulation':'Simulación',
'Puzzle':'Puzzle',
'Casual':'Casual',
'Education':'Educación'
}

df['Género'].replace(map, inplace=True)       # utilizamos un mapeo para traducir los géneros al español

df.head()

df.to_csv('..\\Proyecto_Final\\data\\metacritic_es.csv', index=False)   # guardamos el dataframe en un archivo csv
